お金のUI追加

+++ Assets/ZombiesGame/Player/InventoryProgressSO.cs
@@ -11,7 +11,7 @@ namespace Zombies
         public Action RefleshInventoryEvent;
         
         [Header("お金")]
-        public float NowMoney = 0;
+        public int NowMoney = 0;
         [Header("現在取得している銃の情報")]
         public List<string> GunList = new List<string>();
         [Header("現在アクティブな銃情報")]
